@@ -18,7 +18,7 @@ publish_to: none
 version: 2.0.0
 
 environment:
-  sdk: ">=2.12.1 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   flutter:
