@@ -4,13 +4,13 @@ version: 1.1.1
 homepage: https://github.com/emanuel-braz/one_context.git
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.17.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
   shared_preferences: ^2.0.4
-  back_button_interceptor: ^5.0.0
+  back_button_interceptor: ^6.0.0
 
 dev_dependencies:
   flutter_test:
